% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Says Hello To The User}
\usage{
hello(name)
}
\arguments{
\item{name}{A character string}
}
\value{
A character string that says Hello \code{name}
}
\description{
This packages provides this function to say hello to the user.
}
\examples{
hello(name = "Dusty")
}
