% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycat.R
\name{mycat}
\alias{mycat}
\title{Cat Function}
\usage{
mycat(CatName = "Beat Navy")
}
\arguments{
\item{CatName}{A Cat Name}
}
\value{
ASCII Cat with its name.  An incorrect data type will return an angry cat.
}
\description{
Cat Function
}
\examples{
mycat()
}
