% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randtest.R
\name{randtest}
\alias{randtest}
\title{Randomization Test}
\usage{
randtest(x, y, fun = mean, reps = 10000)
}
\arguments{
\item{x}{double}

\item{y}{double}

\item{fun}{factor}

\item{reps}{double}
}
\value{
a list of probabilities
}
\description{
This fuction is a derivation from the randomization test as
presented in the Kuiper Applied Statistics Text Book.  The
results are in terms of the probability that the null hypothesis
is true
}
