% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randtestvis.R
\name{randtestviz}
\alias{randtestviz}
\title{Randomization Test Visualization}
\usage{
randtestviz(x, y, fun = mean, reps = 10000)
}
\arguments{
\item{x}{double}

\item{y}{double}

\item{fun}{factor}

\item{reps}{double}
}
\value{
A ggplot object
}
\description{
This function returns a `ggplot` that shows the propotion of randomized
x and y vectors (mean(x)-mean(y)) which are greater than the observed
mean(x)-mean(y) vectors.
}
\examples{
#' set.seed(256)
randtest(rnorm(100),rnorm(100,1))
}
